+++ merge.c
@@ -2,23 +2,23 @@
 #include <stdlib.h>
 #include <math.h>
 
-int * slice(int * list, int size, int start, int end) {
-  int* sideList = (int *) malloc(size*sizeof(int));
-  int j = 0;
-  for(int i = start; i <= end; i++){
+long * slice(long * list, long size, long start, long end) {
+  long* sideList = (long *) malloc(size*sizeof(long));
+  long j = 0;
+  for(long i = start; i <= end; i++){
     sideList[j] = list[i];
     j++;
   }
   return sideList;
 }
 
-int * merge(int * list, int start, int middle, int end) {
-  int leftListSize = middle - start;
-  int rightListSize = end - middle + 1;
-  int* leftList = slice(list, leftListSize, start, middle - 1);
-  int* rightList = slice(list, rightListSize, middle, end);
-  int i = 0, j = 0;
-  for(int k = start; k <= end; k++) {
+long * merge(long * list, long start, long middle, long end) {
+  long leftListSize = middle - start;
+  long rightListSize = end - middle + 1;
+  long* leftList = slice(list, leftListSize, start, middle - 1);
+  long* rightList = slice(list, rightListSize, middle, end);
+  long i = 0, j = 0;
+  for(long k = start; k <= end; k++) {
     if(i >= leftListSize) {
       list[k] = rightList[j];
       j++;
@@ -38,9 +38,9 @@ int * merge(int * list, int start, int middle, int end) {
   return list;
 }
 
-void mergeSort(int* list, int start, int end) {
+void mergeSort(long* list, long start, long end) {
   if(end - start >= 1) {
-    int middle = ceil((float)(end + start)/2);
+    long middle = ceil((long double)(end + start)/2);
     mergeSort(list, start, middle - 1);
     mergeSort(list, middle, end);
     merge(list, start, middle, end);
